requirements.txt
Flask==2.2
Werkzeug==2.2.0
Flask-Caching == 2.3.1
python-dotenv == 1.1.0
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.0 